Rename LuaConverter files to lua_converter
