+++ Sprocket/CMakeLists.txt
@@ -81,7 +81,7 @@ add_library(Sprocket STATIC
 
             Scripting/lua_script.cpp
             Scripting/lua_maths.cpp
-            Scripting/LuaConverter.cpp
+            Scripting/lua_converter.cpp
             Scripting/lua_input.cpp)
 
 target_link_libraries(Sprocket PUBLIC
