+++ Sprocket/Scene/lua_ecs.cpp
@@ -2,7 +2,7 @@
 
 #include <Sprocket/Scene/ecs.h>
 #include <Sprocket/Scene/scene.h>
-#include <Sprocket/Scripting/LuaConverter.h>
+#include <Sprocket/Scripting/lua_converter.h>
 #include <Sprocket/Scripting/lua_script.h>
 #include <Sprocket/Core/log.h>
 #include <Sprocket/Utility/Maths.h>
