+++ Sprocket/Scripting/lua_input.cpp
@@ -3,7 +3,7 @@
 #include <Sprocket/Core/log.h>
 #include <Sprocket/Utility/input_store.h>
 #include <Sprocket/Scripting/lua_script.h>
-#include <Sprocket/Scripting/LuaConverter.h>
+#include <Sprocket/Scripting/lua_converter.h>
 
 #include <lua.h>
